+++ src/main/java/com/jian/hobbyadventure/service/ExplorationService.java
@@ -7,8 +7,8 @@ import com.jian.hobbyadventure.domain.Exploration;
 import com.jian.hobbyadventure.domain.UserExploration;
 import com.jian.hobbyadventure.dto.response.ExplorationDetailResponse;
 import com.jian.hobbyadventure.dto.response.ExplorationListItemResponse;
-import com.jian.hobbyadventure.dto.response.PageMeta;
-import com.jian.hobbyadventure.dto.response.PageResponse;
+import com.jian.hobbyadventure.common.response.PageMeta;
+import com.jian.hobbyadventure.common.response.PageResponse;
 import com.jian.hobbyadventure.dto.response.StartExplorationResponse;
 import com.jian.hobbyadventure.repository.CategoryMapper;
 import com.jian.hobbyadventure.repository.ExplorationMapper;
